+++ Base32/ViewController.h
@@ -3,7 +3,7 @@
  *
  * MIT License (MIT)
  *
- * Copyright (c) 2012 Ashkan Farhadtouski
+ * Copyright (c) 2015 Ashkan Farhadtouski
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -26,13 +26,6 @@
 
 #import <UIKit/UIKit.h>
 
-@interface ViewController : UIViewController <UITextViewDelegate>
-
-@property (strong, nonatomic) IBOutlet UITextField *numberToEncodeTextField;
-@property (strong, nonatomic) IBOutlet UITextView *stringToDecodeTextView;
-@property (strong, nonatomic) IBOutlet UITextView *decodedNumberTextView;
-
-- (IBAction)encodeButtonTouchUpInside;
-- (IBAction)decodeButtonTouchUpInside;
+@interface ViewController : UIViewController
 
 @end