+++ RELEASE NOTES.md
@@ -1,3 +1,13 @@
 Version 1.0 (12/29/2012)
 ------------------------
 - Initial release.
+
+Version 1.1 (08/23/2015)
+------------------------
+- Fix a bug in decimal to binary conversion.
+- Fix a bug in character lookup where an invalid index would return a garbage value.
+- Renamed the Base32 class to AFBase32 to avoid potential naming conflicts.
+- Increased the maximum limit to 19.
+- Convert the inline testes to proper XCTesting test cases.
+- Fixed a new inconsistent method return types.
+- Update the rest of the code to the latest coding/configuration practices.